requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo;python_version<"3.9"
base58==2.1.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-ipfs-storage==0.0.4
djangorestframework==3.14.0
drf-spectacular==0.24.2
factory-boy==3.2.1
Faker==15.1.1
flake8==5.0.4
idna==3.4
importlib-resources==5.10.0
inflection==0.5.1
ipfsapi==0.4.4
ipfshttpclient==0.4.13.2
jsonschema==4.16.0
mccabe==0.7.0
mock==4.0.3
multiaddr==0.0.9
netaddr==0.8.0
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pyflakes==2.5.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
requests==2.28.1
responses==0.22.0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
types-toml==0.10.8.1
uritemplate==4.1.1
urllib3==1.26.12
varint==1.0.2
zipp==3.9.0
